Docker-files/Frontend/requirements.txt
Flask==1.0.2
grpcio==1.27.2
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-storage==1.26.0
google-cloud-dataproc==0.7.0